+++ CHANGELOG.md
@@ -87,6 +87,8 @@ and this project adheres to
   - [#4836](https://github.com/bpftrace/bpftrace/pull/4836)
 - Stabilize map declarations (remove 'unstable_map_decl' flag)
   - [#5005](https://github.com/bpftrace/bpftrace/pull/5005)
+- Skip traceable function checks with --unsafe to allow probing functions not listed by ftrace
+  - [#5015](https://github.com/bpftrace/bpftrace/pull/5015)
 #### Deprecated
 - `while` is now deprecated in favor of range-based for loops
   - [#4886](https://github.com/bpftrace/bpftrace/pull/4886)